usb-backend: discard local CD ejection when eject on VM requested
After several experiments I've decided to discard local CD ejection
upon SCSI eject command. First, on Linux unlock/eject operation might
be available only for root. Second, the local CD might be actually not
a local (user may run the client in remote desktop  session). Third,
in general the user should be responsible for presence of media that
it shares and we should not disturb him.