glib: add SpiceQmpPort helper
Add a few helper functions to deal with a QMP port channel, in order
to ease json handling, and wrapping a few commands.

(by convention, the port should have the name
"org.qemu.monitor.qmp.0", but it's not strictly required)

This helper is put into use in the virt-viewer "Add QEMU-like UI: VT
console & basic VM status" series.

Note: this adds a strong dependency on json-glib for
spice-client-glib, a widely available and fairly small
library.

QMP specification is:
https://git.qemu.org/?p=qemu.git;a=blob;f=docs/interop/qmp-spec.txt

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>
Acked-by: Victor Toso <victortoso@redhat.com>