build: Make usb-acl-helper test work under Meson
Meson waits for pipes (stdout and stderr) to be not readable but
our mocking helper does not close these handles causing Meson
to exit with a stack trace.
Note that the fact that the helper is not closing the handles
is a wanted behaviour.

Already reported at https://github.com/mesonbuild/meson/issues/3967.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Eduardo Lima (Etrunko) <etrunko@redhat.com>