display: Trigger wheel scrolling after one 'click'
The scroll wheel on the mouse will trigger a smooth scroll event with a
delta of +/-1.0. The code added in 2212f05 triggers a scroll when the
delta is strictly greater than 1.0. This means that right after
connecting a client, we won't be triggering a scroll with the first
mouse wheel 'click'.
This commit adjusts the check so that we try to scroll when the delta is
equal or greater than 1.0.

https://bugzilla.redhat.com/show_bug.cgi?id=1627823

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>
Acked-by: Marc-André Lureau <marcandre.lureau@gmail.com>