build: Remove need for --enable-static for building tests
Currently, spice-gtk tests are disabled if build of static libraries is
disabled. This commit introduces an intermediate
spice-client-glib-impl.la convenience library which will be used both
for generating the spice-client-glib-2.0 shared library (after applying
the version script), and to link the tests with.
This is better than using --enable-static as we don't need to rebuild
each object twice (static and shared).

This was inspired by similar work Fabiano Fidencio did in libosinfo:
https://www.redhat.com/archives/libosinfo/2018-November/msg00231.html

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>