spice-channel: Avoids spice_channel_read_wire returning success
Although it is really unlikely that errno is <= 0, we should still
make sure spice_channel_read_wire failures are correctly reported in
that event as well.
Such results could end up in buffer overflows.
Moreover, errno is not always set (for instance GLib functions prefer
to set a GError instead or Windows uses WSASetLastError to store
sockets function errors). Sometimes it is also set to an invalid value
to check if some function (like strtod) modifies it.
This is more a safety but better safe than sorry.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>