widget: display a message about required EGL support
Add a new "label" stack page. When EGL is unsupported and a GL scanout
message is received, switch to that page with an appropriate message
hint.

At this point, no further switch back to 2d drawing is implemented,
although it is theorically possible for a server to switch between GL
and 2D, I don't think we need to bother with this corner case.

Alternative implementation also considered:

- a dialog: problematic, the widget may not have an associated
  GtkWindow parent, not very user friendly (one time dialog).

- an info-bar: does not fit well inside display widget, may not be
  well integrated with client application. Can usually be discarded.

- an error signal: would need to be implemented by various client, but
  could give more flexibility on how to present various errors.  Could
  be added later, and replace the "label" page used here by
  default (if a signal handler is detected).

- a property: may be more appropriate, since the error should
  stick. Although it would more complicated to deal with multiple
  errors in that case.

The current solution is not perfect, but should be a decent default,
not requiring client modification. It can be later refined by adding a
way for the client to override the presentation of the message via
signal handlers.

Fixes:
https://gitlab.freedesktop.org/spice/spice-gtk/issues/69

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>
Acked-by: Victor Toso <victortoso@redhat.com>