gst: check pulsesrc version >= 1.15
There is a racy bug in pulsesrc that we can't easily workaround:
https://gitlab.freedesktop.org/gstreamer/gst-plugins-good/merge_requests/69

It will hopefully be fixed with 1.15.

PulseAudio may not be picked by autoaudiosrc, but looking up the
actual source or mimicking GstAutoDetect is unnecessarily complicated.

When pulsesrc < 1.15 is detected, let's drop its rank, so it won't be picked.

Signed-off-by: Marc-André Lureau <marcandre.lureau@redhat.com>
Acked-by: Victor Toso <victortoso@redhat.com>