gstreamer: set timestamp in buffer's GstReferenceTimestampMeta
Currently we set timestamps as buffer's PTS, this value may be changed by
the pipeline in some cases and cause an unexpected buffer warnings (when
GstVideoOverlay is not used). Use GstReferenceTimestampMeta when
synchronization is made by spice.

Before applying this patch you can reproduce the warnings by runing with
DISABLE_GSTVIDEOOVERLAY=1 and starting some audio playback in the guest.

Signed-off-by: Snir Sheriber <ssheribe@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>