channel-display: Instrument code to get frame and queue statistics
This patch is based on some work from Victor Toso (statistics) and
Snir Sheriber (GStreamer probing).
All GstBuffers are queued into decoding_queue and a probe is
attached to the sink in order to understand when the buffers
are decoded.
Previously we didn't add frames to decoding_queue in case of direct
rendering as there was nothing to dequeue them, now we need it to
compute the timing.
When a buffer is decoded the time spent and queue length are
computed.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Victor Toso <victortoso@redhat.com>