display-gst: small refactor get_decoded_frame()
Small refactor to make each code block a bit more obvious.

This code should (1) find the @buffer in the queue; (2) remove all old
elements from queue. That perfectly fit in two loops in sequence, but
they don't need to be nested and they don't need to use the same
pointer (gstframe).

As @num_frames_dropped is only used in the second block, this was
moved as well, together with the debug.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>