gitlab-ci: group and rename jobs
Group by target build instead of command. The focus of each job is to
check any regression for given platform, using 'fedora'/'windows' and
'autotools'/'meson' seems more intuitive.

By doing that we are grouping similar jobs together, this is
intentional as we are reducing the amount of jobs that need to be run
(together with the whole bootstrapping) without losing logs or tests.

Some indentation takes place too, keeping it to 2 spaces as some other
places in the code

v1 -> v2:
- Remove initial '_' from build folder filenames (Frediano);
- Always use '-' instead of '_' in build folder filenames (Frediano);

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Marc-André Lureau <marcandre.lureau@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>