gitlab-ci: move windows dependencies to a variable
To keep all dependencies together. Some (small) effort was made to
distinguish what is necessary for Fedora and what is necessary for
Windows builds in order to install only required packages when job is
executing.

Note that we are adding gettext, gettext-devel and glib2-devel
explicit now on DEPS_COMMON. The reason is that, both Fedora and mingw
builds require some tooling but for Fedora it gets installed
indirectly while for Mingw, it doesn't and configure/build would fail.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>