win-usb-dev: strict comparison of USB devices
If when calling handle_dev_change, the new device has the same
bus:address as an existing device cached in GUdevClient::udev_list, then
an "uevent" signal will not be emitted to notify about the new device.
For example, when due to some reason the 'redirecting' property is set
for a long time and during this time one device is unplugged and then
replugged.
This commit makes the device comparison more strict by checking not only
bus:addr, but also vid:pid.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>
Acked-by: Christophe Fergeau <cfergeau@redhat.com>