Detect timeout conditions more aggressively on Linux
This mitigates a fairly rare problem we see with our kiosk mode clients.
That is, normally if something goes wrong with a client connection
(e.g. the session is killed, or the server is restarted ), the kiosk will
exit on disconnect, and we get a chance to retry the connection, or
present the user with a 'server down' style message.

But in the case of a serious network problem or a server hard power
cycle (i.e. no TCP FIN packets can flow), our end user behavior is not
ideal - the kiosk appears to hang solid, requiring a power cycle.

That's because we've got the stock keepalive timeouts, or about 2 hours
and 11 minutes, before the client sees the disconnect.

This change will cause the client to recognize the server has vanished
without a TCP FIN after 75 seconds.

See this thread:
  https://lists.freedesktop.org/archives/spice-devel/2017-March/036553.html

As well as this bug:
  https://bugzilla.redhat.com/show_bug.cgi?id=1436589

Signed-off-by: Jeremy White <jwhite@codeweavers.com>