channel-usbredir: reduce amount of #ifdef USE_USBREDIR
Not really interesting to keep tracking what is inside or not
USE_USBREDIR on channel-usbredir itself. When usbredir is disabled,
just dummy functions are needed with spice-client.h include.

After this patch, on channel-usbredir.c, we have a single #ifdef
USE_USBREDIR and the #else block is defined in the end of the file.

This patch also adds #ifdef USE_USBREDIR for channel-usbredir-priv.h
as none those functions are defined when usbredir is disabled.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>