gstreamer: Initialize last_mm_time to avoid an overflow
The mm_time is an unsigned 32 bit int but spice_mmtime_diff() returns a
signed 32 bit int. That's reasonable because we normally subtract
frame times which should be at most seconds apart.
But last_mm_time was zero on stream startup, resulting in an overflow
and an uncalled for warning for the first frame in
spice_gst_decoder_queue_frame() if the uptime was greater than about
25 days.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>
Acked-by: Snir Sheriber <ssheribe@redhat.com>