webdav: don't start client on empty message
If the client on the remote side disconnects,
spice-webdav daemon sends a message with no data.

However, if the phodav server already closed
the client connection, client lookup fails and
we open a new connection to phodav, but don't write
anything to it - this should not happen,
so in such case, ignore the message and start
demuxing again.

Signed-off-by: Jakub Janků <jjanku@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>