file-transfer-task: emit signals in main context
Parts of the internal file transfer task API can be invoked in the
coroutine context, so in these cases use g_coroutine_signal_emit and
g_coroutine_object_notify.

Some functions (such as spice_main_channel_clipboard_selection_grab) wake
up the channel and thus can immediately switch to the coroutine context.
Delivering signals in the coroutine context as well, could theoretically
lead to problems with reentrancy, since the user of spice-gtk does not
expect this behavior.
Spice-gtk, on the other hand, assumes that the public functions are called
from the main context. So calling
spice_main_channel_clipboard_selection_grab from the coroutine context
prints an error as we try to switch to a coroutine we are already in.

Such issues are probably not likely, but the fix is easy.

Signed-off-by: Jakub Janků <jjanku@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>