main: migration: check return value to count channels
Although spice_channel_connect() works in idle, if it returns false
it'll not emit any signal further and we would be counting a
'connected' channel that wouldn't be emitting anything.

As other callbacks take this in consideration, we should only
increment the counter if we reached spice-channel::connect_delayed()
callback.

Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>