usb-backend: no emulated isoch devices
SpiceUsbBackendDevice deals with real and emulated devices but there
is no plans to implement emulated isochronous devices.

This patch adds check to edev (emulated device) in the guard, fix the
return value to FALSE instead of plain 0 and return early in case the
code path is around emulated devices.

This is a preparatory patch to extend emulated devices support.

Signed-off-by: Yuri Benditovich <yuri.benditovich@daynix.com>
Signed-off-by: Victor Toso <victortoso@redhat.com>
Acked-by: Frediano Ziglio <fziglio@redhat.com>