implement spice_usb_device_widget_add_err_msg() in usb-device-widget
This is a common function for adding an error status message.

If an old message exists, checks if the new message is not already contained
in the old one. The new message is ignored if it is, added if it's not.

New message string should be dynamically allocated, it's always g_free-ed.

Acked-by: Frediano Ziglio <fziglio@redhat.com>